🧹 Extract PositionRow component and formatters to improve maintainability
