+++ app/page.tsx
@@ -3,23 +3,16 @@
 import { useCallback, useEffect, useMemo, useRef, useState } from "react";
 import { toast } from "sonner";
 
+import { PositionRow } from "@/components/position-row";
+// import PerfBar removed temporarily from "@/components/position-row";
 import { Button } from "@/components/ui/button";
 import { Input } from "@/components/ui/input";
 import { Progress } from "@/components/ui/progress";
 import { Skeleton } from "@/components/ui/skeleton";
 import { type Position, type PositionPerf, useTrancheStore } from "@/lib/store";
+import { currency, budgetNumber, formatSignedPct, perfColor } from "@/lib/utils";
 
-const currency = new Intl.NumberFormat("en-US", {
-  style: "currency",
-  currency: "USD",
-  minimumFractionDigits: 2,
-  maximumFractionDigits: 2,
-});
 
-const budgetNumber = new Intl.NumberFormat("en-US", {
-  minimumFractionDigits: 0,
-  maximumFractionDigits: 2,
-});
 
 const dateFormatter = new Intl.DateTimeFormat("en-US", {
   month: "long",
@@ -39,13 +32,6 @@ function formatEasternDateTime(date: Date) {
   return `${dateFormatter.format(date)} ${timeFormatter.format(date)} EST`;
 }
 
-function formatSignedPct(value: number | null) {
-  if (typeof value !== "number") {
-    return "--";
-  }
-  const sign = value > 0 ? "+" : "";
-  return `${sign}${value.toFixed(2)}%`;
-}
 
 function parseUrlState(search: string): { budget: number | null; positions: Array<{ ticker: string; shares: number }> } | null {
   const params = new URLSearchParams(search);
@@ -91,23 +77,6 @@ function parseUrlState(search: string): { budget: number | null; positions: Arra
   return { budget, positions: shares };
 }
 
-function perfColor(value: number | null) {
-  if (typeof value !== "number") return "text-[#e4e4e7]";
-  return value >= 0 ? "text-[#4ade80]" : "text-[#f87171]";
-}
-
-function PerfBar({ value }: { value: number | null }) {
-  if (typeof value !== "number") {
-    return <div className="h-1.5 w-11 rounded bg-[#27272a]" />;
-  }
-  const width = Math.min(Math.abs(value), 30) / 30;
-  const color = value >= 0 ? "bg-[#4ade80]" : "bg-[#f87171]";
-  return (
-    <div className="h-1.5 w-11 rounded bg-[#27272a]">
-      <div className={`h-full rounded ${color}`} style={{ width: `${Math.max(width * 100, 4)}%` }} />
-    </div>
-  );
-}
 
 
 const BUDGET_WARNING_THRESHOLD = 0.88;
@@ -581,160 +550,33 @@ export default function Home() {
           </div>
 
           <div className="divide-y divide-[#1a1a1e]">
-            {positions.map((position) => {
-              const total = typeof position.price === "number" ? position.price * position.shares : 0;
-              const pctBudget = budget > 0 ? (total / budget) * 100 : 0;
-              const showPopover = activePopoverId === position.id && typeof position.price === "number" && !position.error;
+            {positions.map((position) => (
+              <PositionRow
+                key={position.id}
+                position={position}
+                budget={budget}
+                activePopoverId={activePopoverId}
+                popoverPosition={popoverPosition}
+                perfLoadingMap={perfLoadingMap}
+                tickerInputRef={(node) => {
+                  tickerInputRefs.current[position.id] = node;
+                }}
+                shareInputRef={(node) => {
+                  shareInputRefs.current[position.id] = node;
+                }}
+                handleTickerChange={handleTickerChange}
+                handleTickerKeyDown={handleTickerKeyDown}
+                handleTickerBlur={handleTickerBlur}
+                handlePositionMouseEnter={handlePositionMouseEnter}
+                handlePositionMouseLeave={handlePositionMouseLeave}
+                handlePopoverMouseEnter={handlePopoverMouseEnter}
+                handlePopoverMouseLeave={handlePopoverMouseLeave}
+                incrementShares={incrementShares}
+                handleShareChange={handleShareChange}
+                removePosition={removePosition}
 
-              const perfRows: Array<{ label: string; value: number | null; loading: boolean }> = [
-                { label: "1D", value: position.changePct1D, loading: false },
-                {
-                  label: "1W",
-                  value: position.perf?.changePct1W ?? null,
-                  loading: !!perfLoadingMap[position.id] && !position.perf,
-                },
-                {
-                  label: "3M",
-                  value: position.perf?.changePct3M ?? null,
-                  loading: !!perfLoadingMap[position.id] && !position.perf,
-                },
-                {
-                  label: "YTD",
-                  value: position.perf?.changePctYTD ?? null,
-                  loading: !!perfLoadingMap[position.id] && !position.perf,
-                },
-                {
-                  label: "1Y",
-                  value: position.perf?.changePct1Y ?? null,
-                  loading: !!perfLoadingMap[position.id] && !position.perf,
-                },
-              ];
-
-              return (
-                <div
-                  key={position.id}
-                  className="grid min-w-[740px] grid-cols-[88px_minmax(280px,1fr)_148px_112px_88px_36px] items-center gap-3 px-3 py-3 sm:px-4"
-                >
-                  <Input
-                    ref={(node) => {
-                      tickerInputRefs.current[position.id] = node;
-                    }}
-                    value={position.ticker}
-                    maxLength={10}
-                    placeholder="TICK"
-                    onChange={(event) => handleTickerChange(position.id, event.target.value)}
-                    onKeyDown={(event) => handleTickerKeyDown(event, position.id, position.ticker)}
-                    onBlur={() => handleTickerBlur(position.id, position.ticker)}
-                    className="h-9 border-[#27272a] bg-[#09090b] px-2 text-center text-base font-semibold uppercase text-[#f59e0b] [font-family:var(--font-mono)] placeholder:text-[#3f3f46]"
-                  />
-
-                  <div
-                    className="relative"
-                    onMouseEnter={(event) => handlePositionMouseEnter(position, event.currentTarget)}
-                    onMouseLeave={handlePositionMouseLeave}
-                  >
-                    {position.loading ? (
-                      <div className="space-y-2 py-0.5">
-                        <Skeleton className="h-3 w-44 bg-[#27272a]" />
-                        <Skeleton className="h-3 w-28 bg-[#27272a]" />
-                      </div>
-                    ) : position.error ? (
-                      <p className="text-sm text-[#f87171]">{position.error}</p>
-                    ) : (
-                      <div className="flex items-center gap-2">
-                        <div className="min-w-0">
-                          <p className="truncate text-xs text-[#71717a]">
-                            {position.name || (position.ticker ? "Waiting for quote" : "Enter ticker")}
-                          </p>
-                          {typeof position.price === "number" ? (
-                            <p className="mt-0.5 flex items-center gap-2 text-sm [font-family:var(--font-mono)]">
-                              <span className="text-[#4ade80]">{currency.format(position.price)}</span>
-                              <span className={perfColor(position.changePct1D)}>{formatSignedPct(position.changePct1D)}</span>
-                              <span className="text-[#52525b]">···</span>
-                            </p>
-                          ) : (
-                            <p className="mt-0.5 text-sm text-[#52525b] [font-family:var(--font-mono)]">--</p>
-                          )}
-                        </div>
-                      </div>
-                    )}
-
-                    {showPopover && popoverPosition && (
-                      <div
-                        className="fixed z-40 w-64 rounded-sm border border-[#27272a] bg-[#121214] p-3 shadow-xl"
-                        style={{ top: popoverPosition.top, left: popoverPosition.left }}
-                        onMouseEnter={handlePopoverMouseEnter}
-                        onMouseLeave={handlePopoverMouseLeave}
-                      >
-                        <p className="text-lg text-[#e4e4e7] [font-family:var(--font-mono)]">
-                          {currency.format(position.price ?? 0)}
-                        </p>
-                        <p className={`mt-1 text-sm ${perfColor(position.changePct1D)}`}>
-                          {typeof position.changePct1D === "number"
-                            ? `${position.changePct1D >= 0 ? "▲" : "▼"} ${formatSignedPct(position.changePct1D)} today`
-                            : "-- today"}
-                        </p>
-                        <div className="mt-3 space-y-2">
-                          {perfRows.map((row) => (
-                            <div key={row.label} className="grid grid-cols-[24px_44px_1fr] items-center gap-2 text-xs">
-                              <span className="text-[#a1a1aa]">{row.label}</span>
-                              {row.loading ? <Skeleton className="h-1.5 w-11 bg-[#27272a]" /> : <PerfBar value={row.value} />}
-                              <span className={`text-right [font-family:var(--font-mono)] ${perfColor(row.value)}`}>
-                                {row.loading ? <Skeleton className="ml-auto h-3 w-12 bg-[#27272a]" /> : formatSignedPct(row.value)}
-                              </span>
-                            </div>
-                          ))}
-                        </div>
-                        <p className="mt-3 text-[11px] text-[#52525b]">Prices may be delayed</p>
-                      </div>
-                    )}
-                  </div>
-
-                  <div className="grid grid-cols-[32px_1fr_32px] items-center gap-1">
-                    <Button
-                      variant="outline"
-                      size="icon-sm"
-                      onClick={(event) => incrementShares(position, event, -1)}
-                      className="h-8 w-8 border-[#27272a] bg-[#09090b] text-[#e4e4e7] hover:bg-[#202024]"
-                    >
-                      -
-                    </Button>
-                    <Input
-                      ref={(node) => {
-                        shareInputRefs.current[position.id] = node;
-                      }}
-                      value={position.shares}
-                      onChange={(event) => handleShareChange(event, position.id)}
-                      className="h-8 border-[#27272a] bg-[#09090b] px-2 text-center text-sm [font-family:var(--font-mono)]"
-                    />
-                    <Button
-                      variant="outline"
-                      size="icon-sm"
-                      onClick={(event) => incrementShares(position, event, 1)}
-                      className="h-8 w-8 border-[#27272a] bg-[#09090b] text-[#e4e4e7] hover:bg-[#202024]"
-                    >
-                      +
-                    </Button>
-                  </div>
-
-                  <p className="text-right text-sm [font-family:var(--font-mono)]">{currency.format(total)}</p>
-                  <div className="flex justify-center">
-                    <span className="rounded border border-[#7c5412] bg-[#2f230f] px-2 py-0.5 text-xs text-[#f59e0b] [font-family:var(--font-mono)]">
-                      {pctBudget.toFixed(1)}%
-                    </span>
-                  </div>
-                  <Button
-                    variant="ghost"
-                    size="icon-sm"
-                    onClick={() => removePosition(position.id)}
-                    className="h-8 w-8 text-[#71717a] hover:bg-[#202024] hover:text-[#f87171]"
-                    aria-label="Remove position"
-                  >
-                    ×
-                  </Button>
-                </div>
-              );
-            })}
+              />
+            ))}
           </div>
 
             <div className="p-4">
